Rename expect -> console
